LayoutTests/platform/chromium/http/tests/security/dataURL/xss-DENIED-to-data-url-in-foreign-domain-subframe-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL data:text/html,<html><head><script>function loaded() {if (window.layoutTestController)layoutTestController.globalFlag = true;}</script></head><body onload='loaded();'><p id='accessMe'>PASS: Cross frame access from a frame on a foreign domain denied!</p><p>Inner-inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame that the main frame is trying to access.  It should not have access to it.</p></body></html> from frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-to-data-url-in-foreign-domain-subframe.html. Domains, protocols and ports must match.

The scenario for this test is that you have an iframe with content from a foreign domain. In that foreign content is an iframe which loads a data: URL. This tests that this main document does not have access to that data: URL loaded iframe.


PASS: Cross frame access to a data: URL embed in a frame on a foreign domain denied!


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Inner iframe on a foreign domain.



--------
Frame: 'aFrame'
--------
PASS: Cross frame access from a frame on a foreign domain denied!

Inner-inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame that the main frame is trying to access. It should not have access to it.
